gpu-api/requirements.txt
Flask==2.1.1
gunicorn==20.1.0
gpustat==0.6.0
torchvision